Explicitly check against urlNames
To avoid try/catch.
